Support private channels which don't leak info
Currently, we can hide a channel entirely from seen reporting, but there was no
way to still be aware that someone had been talking, without their messages in
that channel leaking out into another channel.

This commit adds new "privatechan" / "unprivatechan" commands which mark
channels as "private"; if a seen response is from a private channel, and the
seen request came from a different channel, then the contents of their message
will not be disclosed in the response.

In other words, it's intended for channels we don't need to hide the existence
of, but do need to ensure the contents of discussion in them aren't leaked into
less privileged channels.